more internals renaming
